requirements.txt
flask>=3.0
lseg-data>=2.0
pandas>=2.0
numpy>=1.26
pandas-market-calendars>=4.4
pytz>=2024.1
gunicorn>=22.0
openpyxl>=3.1
